requirements.txt
beautifulsoup4~=4.12.2
black==22.3.0
gunicorn==20.1.0
requests==2.27.1
lxml==4.9.2
requests-cache==0.9.4
grobid-client==0.8.3
pyalex~=0.7
tenacity~=8.2.1
aioboto3~=11.2.0
aiobotocore~=2.5.0
fastapi~=0.101.0
uvicorn~=0.23.2
starlette~=0.27.0
pdfkit~=1.0.0